+++ aiograpi/mixins/timeline.py
@@ -44,6 +44,23 @@ class ReelsMixin(ClientMixin):
         """
         return await self.reels_timeline_media("explore_reels", amount, last_media_pk)
 
+    async def friends_reels(self, amount: int = 10, last_media_pk: int = 0) -> List[Media]:
+        """
+        Get Friends tab reels media
+
+        Parameters
+        ----------
+        amount: int, optional
+            Maximum number of media to return, default is 10
+        last_media_pk: int, optional
+            Last PK user has seen, function will return medias after this pk. Default is 0
+        Returns
+        -------
+        List[Media]
+            A list of objects of Media
+        """
+        return await self.reels_timeline_media("friends_reels", amount, last_media_pk)
+
     async def reels_timeline_media(self, collection_pk: str, amount: int = 10, last_media_pk: int = 0) -> List[Media]:
         """
         Get reels timeline media in a collection
@@ -66,6 +83,7 @@ class ReelsMixin(ClientMixin):
         private_request_endpoint = {
             "reels": "clips/connected/",
             "explore_reels": "clips/discover/",
+            "friends_reels": "clips/discover/social/",
         }.get(collection_pk)
         if not private_request_endpoint:
             self.logger.warning("Unsupported reels timeline collection: %r", collection_pk)